Simplify and refactor tests
